Bug 794553 - Be more selective about uncommitted changes in versionbump script;r=jhammel
We were being both too permissive and not permissive enough. We were complaining
about changes which could not cause problems (untracked files that have nothing
to do with the modules we're version bumping) and NOT complaining about files
that could (ignored files that would get inadvertently uploaded to pypi).

This new patch only checks for uncommitted files in the directories of modules
we're updating, but does the check more thoroughly for ignored files as well.