Bug 800655 - Make dmADB's pushFile work like dmSUT's;r=mcote
dmSUT's pushFile will throw an exception if you attempt to push a file
to a directory (e.g. pushing the file "foo" to "/mnt/sdcard"). With dmADB,
we put the file in the directory if not using run-as, and would silently fail
if we *were* using run-as. This patch attempts to make dmADB do what dmSUT
does, which is throw an exception if we attempt to directly use pushFile with
a directory as a destination.